moving decoder into wsmv module
