+++ tests/spec/wsmv/command_output_decoder_spec.rb
@@ -1,8 +1,8 @@
 # encoding: UTF-8
 
-require 'winrm/command_output_decoder'
+require 'winrm/wsmv/command_output_decoder'
 
-describe WinRM::CommandOutputDecoder, unit: true do
+describe WinRM::WSMV::CommandOutputDecoder, unit: true do
   let(:raw_output_with_bom) do
     '77u/' \
     'ICAgQ29ubmVjdGlvbi1zcGVjaWZpYyBETlMgU3VmZml4ICAuIDogDQogICBMaW5rLWxvY2FsIElQdjYgQWRkcmVzcyA' \
